change gradio server to just a tab in multi-view
