--- fastchat/serve/gradio_web_server_vision.py
@@ -1,646 +1,0 @@
-"""
-The gradio demo server for chatting with a single model.
-"""
-
-import argparse
-from collections import defaultdict
-import datetime
-import json
-import os
-import random
-import time
-import uuid
-
-import gradio as gr
-import requests
-
-from fastchat.conversation import SeparatorStyle
-from fastchat.constants import (
-    LOGDIR,
-    WORKER_API_TIMEOUT,
-    ErrorCode,
-    MODERATION_MSG,
-    CONVERSATION_LIMIT_MSG,
-    SERVER_ERROR_MSG,
-    INACTIVE_MSG,
-    INPUT_CHAR_LEN_LIMIT,
-    CONVERSATION_TURN_LIMIT,
-    SESSION_EXPIRATION_TIME,
-)
-from fastchat.serve.gradio_web_server import (
-    State,
-    set_global_vars,
-    get_conv_log_filename,
-    get_model_list,
-    load_demo_single,
-    load_demo,
-    upvote_last_response,
-    downvote_last_response,
-    flag_last_response,
-    post_process_code,
-    get_model_description_md,
-    headers,
-    no_change_btn,
-    enable_btn,
-    disable_btn,
-    enable_moderation,
-    acknowledgment_md,
-    block_css,
-    controller_url,
-)
-from fastchat.model.llava.constants import LLAVA_IMAGE_TOKEN
-from fastchat.model.model_adapter import get_conversation_template
-from fastchat.model.model_registry import get_model_info, model_info
-from fastchat.serve.api_provider import (
-    openai_api_stream_iter,
-)
-from fastchat.utils import (
-    build_logger,
-    oai_moderation,
-    get_window_url_params_js,
-    get_window_url_params_with_tos_js,
-    parse_gradio_auth_creds,
-)
-
-
-logger = build_logger("gradio_web_server", "gradio_web_server.log")
-ip_expiration_dict = defaultdict(lambda: 0)
-
-# Information about custom OpenAI compatible API models.
-# JSON file format:
-# {
-#     "vicuna-7b": {
-#         "model_name": "vicuna-7b-v1.5",
-#         "api_base": "http://8.8.8.55:5555/v1",
-#         "api_key": "password"
-#     },
-# }
-openai_compatible_models_info = {}
-
-
-def set_global_vars(controller_url_, enable_moderation_):
-    global controller_url, enable_moderation
-    controller_url = controller_url_
-    enable_moderation = enable_moderation_
-
-
-def regenerate(state, request: gr.Request):
-    logger.info(f"regenerate. ip: {request.client.host}")
-    state.conv.update_last_message(None)
-    return (state, state.to_gradio_chatbot(), "", None) + (disable_btn,) * 5
-
-
-def clear_history(request: gr.Request):
-    logger.info(f"clear_history. ip: {request.client.host}")
-    state = None
-    return (state, [], "", None) + (disable_btn,) * 5
-
-
-def add_text(state, model_selector, text, image, request: gr.Request):
-    ip = request.client.host
-    logger.info(f"add_text. ip: {ip}. len: {len(text)}")
-
-    if state is None:
-        state = State(model_selector)
-
-    if len(text) <= 0:
-        state.skip_next = True
-        return (state, state.to_gradio_chatbot(), "", None) + (no_change_btn,) * 5
-
-    if ip_expiration_dict[ip] < time.time():
-        logger.info(f"inactive. ip: {request.client.host}. text: {text}")
-        state.skip_next = True
-        return (state, state.to_gradio_chatbot(), INACTIVE_MSG, None) + (
-            no_change_btn,
-        ) * 5
-
-    if enable_moderation:
-        flagged = oai_moderation(text)
-        if flagged:
-            logger.info(f"violate moderation. ip: {request.client.host}. text: {text}")
-            state.skip_next = True
-            return (state, state.to_gradio_chatbot(), MODERATION_MSG, None) + (
-                no_change_btn,
-            ) * 5
-
-    if image is not None and len(state.conv.get_images()) > 0:
-        # reset convo with new image
-        state.conv = get_conversation_template(state.model_name)
-
-    conv = state.conv
-    if (len(conv.messages) - conv.offset) // 2 >= CONVERSATION_TURN_LIMIT:
-        logger.info(f"conversation turn limit. ip: {request.client.host}. text: {text}")
-        state.skip_next = True
-        return (state, state.to_gradio_chatbot(), CONVERSATION_LIMIT_MSG, None) + (
-            no_change_btn,
-        ) * 5
-
-    text = text[:INPUT_CHAR_LEN_LIMIT]  # Hard cut-off
-
-    if image is not None:
-        if "llava" in state.model_name.lower():
-            if LLAVA_IMAGE_TOKEN not in text:
-                text = text + "\n" + LLAVA_IMAGE_TOKEN
-
-        text = (
-            text,
-            image,
-        )
-
-    conv.append_message(conv.roles[0], text)
-    conv.append_message(conv.roles[1], None)
-    return (state, state.to_gradio_chatbot(), "", None) + (disable_btn,) * 5
-
-
-def load_demo(url_params, request: gr.Request):
-    global models
-
-    ip = request.client.host
-    logger.info(f"load_demo. ip: {ip}. params: {url_params}")
-    ip_expiration_dict[ip] = time.time() + SESSION_EXPIRATION_TIME
-
-    if args.model_list_mode == "reload":
-        models = get_model_list(
-            controller_url,
-            args.register_openai_compatible_models,
-            args.add_chatgpt,
-            args.add_claude,
-            args.add_palm,
-        )
-
-    return load_demo_single(models, url_params)
-
-
-def model_worker_stream_iter(
-    conv,
-    model_name,
-    worker_addr,
-    prompt,
-    images,
-    temperature,
-    repetition_penalty,
-    top_p,
-    max_new_tokens,
-):
-    # Make requests
-    if len(images) > 0:
-        gen_params = {
-            "model": model_name,
-            "prompt": prompt,
-            "images": images,
-            "temperature": temperature,
-            "repetition_penalty": repetition_penalty,
-            "top_p": top_p,
-            "max_new_tokens": max_new_tokens,
-            "stop": conv.stop_str,
-            "stop_token_ids": conv.stop_token_ids,
-            "echo": False,
-        }
-    else:
-        gen_params = {
-            "model": model_name,
-            "prompt": prompt,
-            "temperature": temperature,
-            "repetition_penalty": repetition_penalty,
-            "top_p": top_p,
-            "max_new_tokens": max_new_tokens,
-            "stop": conv.stop_str,
-            "stop_token_ids": conv.stop_token_ids,
-            "echo": False,
-        }
-    logger.info(f"==== request ====\n{gen_params}")
-
-    # Stream output
-    response = requests.post(
-        worker_addr + "/worker_generate_stream",
-        headers=headers,
-        json=gen_params,
-        stream=True,
-        timeout=WORKER_API_TIMEOUT,
-    )
-    for chunk in response.iter_lines(decode_unicode=False, delimiter=b"\0"):
-        if chunk:
-            data = json.loads(chunk.decode())
-            yield data
-
-
-def bot_response(state, temperature, top_p, max_new_tokens, request: gr.Request):
-    logger.info(f"bot_response. ip: {request.client.host}")
-    start_tstamp = time.time()
-    temperature = float(temperature)
-    top_p = float(top_p)
-    max_new_tokens = int(max_new_tokens)
-
-    if state.skip_next:
-        # This generate call is skipped due to invalid inputs
-        state.skip_next = False
-        yield (state, state.to_gradio_chatbot()) + (no_change_btn,) * 5
-        return
-
-    conv, model_name = state.conv, state.model_name
-    if model_name in openai_compatible_models_info:
-        model_info = openai_compatible_models_info[model_name]
-        prompt = conv.to_openai_api_messages()
-        stream_iter = openai_api_stream_iter(
-            model_info["model_name"],
-            prompt,
-            temperature,
-            top_p,
-            max_new_tokens,
-            api_base=model_info["api_base"],
-            api_key=model_info["api_key"],
-        )
-    else:
-        # Query worker address
-        ret = requests.post(
-            controller_url + "/get_worker_address", json={"model": model_name}
-        )
-        worker_addr = ret.json()["address"]
-        logger.info(f"model_name: {model_name}, worker_addr: {worker_addr}")
-
-        # No available worker
-        if worker_addr == "":
-            conv.update_last_message(SERVER_ERROR_MSG)
-            yield (
-                state,
-                state.to_gradio_chatbot(),
-                disable_btn,
-                disable_btn,
-                disable_btn,
-                enable_btn,
-                enable_btn,
-            )
-            return
-
-        # Construct prompt.
-        # We need to call it here, so it will not be affected by "▌".
-        prompt = conv.get_prompt()
-        images = conv.get_images()
-
-        # Set repetition_penalty
-        repetition_penalty = 1.0
-
-        stream_iter = model_worker_stream_iter(
-            conv,
-            model_name,
-            worker_addr,
-            prompt,
-            images,
-            temperature,
-            repetition_penalty,
-            top_p,
-            max_new_tokens,
-        )
-
-    conv.update_last_message("▌")
-    yield (state, state.to_gradio_chatbot()) + (disable_btn,) * 5
-
-    try:
-        for i, data in enumerate(stream_iter):
-            if data["error_code"] == 0:
-                if i % 8 != 0:  # reduce gradio's overhead
-                    continue
-                output = data["text"].strip()
-                conv.update_last_message(output + "▌")
-                yield (state, state.to_gradio_chatbot()) + (disable_btn,) * 5
-            else:
-                output = data["text"] + f"\n\n(error_code: {data['error_code']})"
-                conv.update_last_message(output)
-                yield (state, state.to_gradio_chatbot()) + (
-                    disable_btn,
-                    disable_btn,
-                    disable_btn,
-                    enable_btn,
-                    enable_btn,
-                )
-                return
-        output = data["text"].strip()
-        if "vicuna" in model_name:
-            output = post_process_code(output)
-        conv.update_last_message(output)
-        yield (state, state.to_gradio_chatbot()) + (enable_btn,) * 5
-    except requests.exceptions.RequestException as e:
-        conv.update_last_message(
-            f"{SERVER_ERROR_MSG}\n\n"
-            f"(error_code: {ErrorCode.GRADIO_REQUEST_ERROR}, {e})"
-        )
-        yield (state, state.to_gradio_chatbot()) + (
-            disable_btn,
-            disable_btn,
-            disable_btn,
-            enable_btn,
-            enable_btn,
-        )
-        return
-    except Exception as e:
-        conv.update_last_message(
-            f"{SERVER_ERROR_MSG}\n\n"
-            f"(error_code: {ErrorCode.GRADIO_STREAM_UNKNOWN_ERROR}, {e})"
-        )
-        yield (state, state.to_gradio_chatbot()) + (
-            disable_btn,
-            disable_btn,
-            disable_btn,
-            enable_btn,
-            enable_btn,
-        )
-        return
-
-    finish_tstamp = time.time()
-    logger.info(f"{output}")
-
-    with open(get_conv_log_filename(), "a") as fout:
-        data = {
-            "tstamp": round(finish_tstamp, 4),
-            "type": "chat",
-            "model": model_name,
-            "gen_params": {
-                "temperature": temperature,
-                "top_p": top_p,
-                "max_new_tokens": max_new_tokens,
-            },
-            "start": round(start_tstamp, 4),
-            "finish": round(finish_tstamp, 4),
-            "state": state.dict(),
-            "ip": request.client.host,
-        }
-        fout.write(json.dumps(data) + "\n")
-
-
-def build_single_vision_language_model_ui(models, add_promotion_links=False):
-    promotion = (
-        """
-- | [GitHub](https://github.com/lm-sys/FastChat) | [Dataset](https://github.com/lm-sys/FastChat/blob/main/docs/dataset_release.md) | [Twitter](https://twitter.com/lmsysorg) | [Discord](https://discord.gg/HSWAKCrnFx) |
-- Introducing Llama 2: The Next Generation Open Source Large Language Model. [[Website]](https://ai.meta.com/llama/)
-- Vicuna: An Open-Source Chatbot Impressing GPT-4 with 90% ChatGPT Quality. [[Blog]](https://lmsys.org/blog/2023-03-30-vicuna/)
-"""
-        if add_promotion_links
-        else ""
-    )
-
-    notice_markdown = f"""
-# 🏔️ Chat with Open Large Vision-Language Models
-{promotion}
-
-### Choose a model to chat with
-"""
-
-    state = gr.State()
-    model_description_md = get_model_description_md(models)
-    gr.Markdown(notice_markdown + model_description_md, elem_id="notice_markdown")
-
-    with gr.Row():
-        with gr.Column(scale=3):
-            with gr.Row(elem_id="model_selector_row"):
-                model_selector = gr.Dropdown(
-                    choices=models,
-                    value=models[0] if len(models) > 0 else "",
-                    interactive=True,
-                    show_label=False,
-                    container=False,
-                )
-
-            imagebox = gr.Image(type="pil")
-
-            cur_dir = os.path.dirname(os.path.abspath(__file__))
-
-            with gr.Accordion("Parameters", open=False) as parameter_row:
-                temperature = gr.Slider(
-                    minimum=0.0,
-                    maximum=1.0,
-                    value=0.2,
-                    step=0.1,
-                    interactive=True,
-                    label="Temperature",
-                )
-                top_p = gr.Slider(
-                    minimum=0.0,
-                    maximum=1.0,
-                    value=0.7,
-                    step=0.1,
-                    interactive=True,
-                    label="Top P",
-                )
-                max_output_tokens = gr.Slider(
-                    minimum=0,
-                    maximum=1024,
-                    value=512,
-                    step=64,
-                    interactive=True,
-                    label="Max output tokens",
-                )
-
-        with gr.Column(scale=8):
-            chatbot = gr.Chatbot(
-                elem_id="chatbot", label="Scroll down and start chatting", height=550
-            )
-
-            with gr.Row():
-                with gr.Column(scale=8):
-                    textbox = gr.Textbox(
-                        show_label=False,
-                        placeholder="Enter your prompt here and press ENTER",
-                        container=False,
-                        elem_id="input_box",
-                    )
-                with gr.Column(scale=1, min_width=50):
-                    send_btn = gr.Button(value="Send", variant="primary")
-            with gr.Row(elem_id="buttons"):
-                upvote_btn = gr.Button(value="👍  Upvote", interactive=False)
-                downvote_btn = gr.Button(value="👎  Downvote", interactive=False)
-                flag_btn = gr.Button(value="⚠️  Flag", interactive=False)
-                regenerate_btn = gr.Button(value="🔄  Regenerate", interactive=False)
-                clear_btn = gr.Button(value="🗑️  Clear", interactive=False)
-
-            gr.Examples(
-                examples=[
-                    [
-                        f"{cur_dir}/example_images/dog.jpeg",
-                        "What animal is in this photo?",
-                    ],
-                    [
-                        f"{cur_dir}/example_images/sunset.jpg",
-                        "Where was this picture taken?",
-                    ],
-                ],
-                inputs=[imagebox, textbox],
-            )
-
-    if add_promotion_links:
-        gr.Markdown(acknowledgment_md)
-
-    # Register listeners
-    btn_list = [upvote_btn, downvote_btn, flag_btn, regenerate_btn, clear_btn]
-    upvote_btn.click(
-        upvote_last_response,
-        [state, model_selector],
-        [textbox, upvote_btn, downvote_btn, flag_btn],
-    )
-    downvote_btn.click(
-        downvote_last_response,
-        [state, model_selector],
-        [textbox, upvote_btn, downvote_btn, flag_btn],
-    )
-    flag_btn.click(
-        flag_last_response,
-        [state, model_selector],
-        [textbox, upvote_btn, downvote_btn, flag_btn],
-    )
-    regenerate_btn.click(
-        regenerate, state, [state, chatbot, textbox, imagebox] + btn_list
-    ).then(
-        bot_response,
-        [state, temperature, top_p, max_output_tokens],
-        [state, chatbot] + btn_list,
-    )
-    clear_btn.click(clear_history, None, [state, chatbot, textbox, imagebox] + btn_list)
-
-    model_selector.change(
-        clear_history, None, [state, chatbot, textbox, imagebox] + btn_list
-    )
-
-    textbox.submit(
-        add_text,
-        [state, model_selector, textbox, imagebox],
-        [state, chatbot, textbox, imagebox] + btn_list,
-    ).then(
-        bot_response,
-        [state, temperature, top_p, max_output_tokens],
-        [state, chatbot] + btn_list,
-    )
-    send_btn.click(
-        add_text,
-        [state, model_selector, textbox, imagebox],
-        [state, chatbot, textbox, imagebox] + btn_list,
-    ).then(
-        bot_response,
-        [state, temperature, top_p, max_output_tokens],
-        [state, chatbot] + btn_list,
-    )
-
-    return [state, model_selector]
-
-
-def build_demo(models):
-    with gr.Blocks(
-        title="Chat with Open Large Vision-Language Models",
-        theme=gr.themes.Default(),
-        css=block_css,
-    ) as demo:
-        url_params = gr.JSON(visible=False)
-
-        state, model_selector = build_single_vision_language_model_ui(models)
-
-        if args.model_list_mode not in ["once", "reload"]:
-            raise ValueError(f"Unknown model list mode: {args.model_list_mode}")
-
-        if args.show_terms_of_use:
-            load_js = get_window_url_params_with_tos_js
-        else:
-            load_js = get_window_url_params_js
-
-        demo.load(
-            load_demo,
-            [url_params],
-            [
-                state,
-                model_selector,
-            ],
-            _js=load_js,
-        )
-
-    return demo
-
-
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser()
-    parser.add_argument("--host", type=str, default="0.0.0.0")
-    parser.add_argument("--port", type=int)
-    parser.add_argument(
-        "--share",
-        action="store_true",
-        help="Whether to generate a public, shareable link",
-    )
-    parser.add_argument(
-        "--controller-url",
-        type=str,
-        default="http://localhost:21001",
-        help="The address of the controller",
-    )
-    parser.add_argument(
-        "--concurrency-count",
-        type=int,
-        default=10,
-        help="The concurrency count of the gradio queue",
-    )
-    parser.add_argument(
-        "--model-list-mode",
-        type=str,
-        default="once",
-        choices=["once", "reload"],
-        help="Whether to load the model list once or reload the model list every time",
-    )
-    parser.add_argument(
-        "--moderate",
-        action="store_true",
-        help="Enable content moderation to block unsafe inputs",
-    )
-    parser.add_argument(
-        "--show-terms-of-use",
-        action="store_true",
-        help="Shows term of use before loading the demo",
-    )
-    parser.add_argument(
-        "--add-chatgpt",
-        action="store_true",
-        help="Add OpenAI's ChatGPT models (gpt-3.5-turbo, gpt-4)",
-    )
-    parser.add_argument(
-        "--add-claude",
-        action="store_true",
-        help="Add Anthropic's Claude models (claude-2, claude-instant-1)",
-    )
-    parser.add_argument(
-        "--add-palm",
-        action="store_true",
-        help="Add Google's PaLM model (PaLM 2 for Chat: chat-bison@001)",
-    )
-    parser.add_argument(
-        "--register-openai-compatible-models",
-        type=str,
-        help="Register custom OpenAI API compatible models by loading them from a JSON file",
-    )
-    parser.add_argument(
-        "--gradio-auth-path",
-        type=str,
-        help='Set the gradio authentication file path. The file should contain one or more user:password pairs in this format: "u1:p1,u2:p2,u3:p3"',
-    )
-    args = parser.parse_args()
-    logger.info(f"args: {args}")
-
-    # Set global variables
-    set_global_vars(args.controller_url, args.moderate)
-    models = get_model_list(
-        args.controller_url,
-        args.register_openai_compatible_models,
-        args.add_chatgpt,
-        args.add_claude,
-        args.add_palm,
-    )
-
-    # Set authorization credentials
-    auth = None
-    if args.gradio_auth_path is not None:
-        auth = parse_gradio_auth_creds(args.gradio_auth_path)
-
-    # Launch the demo
-    demo = build_demo(models)
-    demo.queue(
-        concurrency_count=args.concurrency_count, status_update_rate=10, api_open=False
-    ).launch(
-        server_name=args.host,
-        server_port=args.port,
-        share=args.share,
-        max_threads=200,
-        auth=auth,
-    )
